Add files via upload
This GitHub repository contains the source code of a magnificent website dedicated to the sale of flowers. Our platform offers a unique online floral experience, allowing customers to browse and buy a variety of fresh and vibrant flowers. With user-friendly navigation, ordering and delivery features, our site aims to bring the beauty of nature directly to our customers.